Scripts/Clinical_GraphRAG_Chatbot/requirements.txt
streamlit>=1.28.0
neo4j>=5.14.0
pymilvus>=2.3.0
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
pyvis>=0.3.2

